Move suppression test to core as suppression was moved to internal and will become core
