+++ detekt-core/src/test/resources/suppression/SuppressedObject.kt
